+++ sancho-src/src/sancho/core/MLDonkeyCoreMonitor.java
@@ -4,14 +4,14 @@ import java.net.Socket;
 import sancho.model.mldonkey.utility.MessageBuffer;
 
 public class MLDonkeyCoreMonitor extends MLDonkeyCore {
-   public MLDonkeyCoreMonitor(Socket var1, String var2, String var3, boolean var4) {
-      super(var1, var2, var3, var4);
+   public MLDonkeyCoreMonitor(Socket socket, String username, String password, boolean pollMode) {
+      super(socket, username, password, pollMode);
    }
 
-   void processMessage(int var1, MessageBuffer var2) {
-      switch (var1) {
+   void processMessage(int opcode, MessageBuffer buffer) {
+      switch (opcode) {
          case 0:
-            this.readCoreProtocol(var2);
+            this.readCoreProtocol(buffer);
             break;
          case 47:
             this.disconnect();
@@ -22,7 +22,7 @@ public class MLDonkeyCoreMonitor extends MLDonkeyCore {
                this.notifyInitialized();
             }
 
-            this.getClientStats().read(var2);
+            this.getClientStats().read(buffer);
       }
    }
 }